+++ src/Service/Email/EmailServiceInterface.php
@@ -2,6 +2,8 @@
 
 namespace App\Service\Email;
 
+use App\Entity\User;
+
 interface EmailServiceInterface
 {
     /**
@@ -10,4 +12,6 @@ interface EmailServiceInterface
      * @throws EmailNotSendException
      */
     public function sendBatch(array $emailsData): void;
+
+    public function sendInitPassword(User $user, string $token);
 }